refactor(wave-1): area shoelace + overlay-from-polygons + transient brush raster
Consolidated merge of the three wave-1 sessions per knowledge/030. Written
directly on master (no worktrees) after the parallel-worktree harness
produced a mix of uncommitted partial work.

Wave 1A — compute_area_rows now calls masking.polygon_area(vertices) rather
than reading state.region_areas; AreaRow.area_px becomes float; CSV
fixtures regenerated.

Wave 1B — ImageCanvas._update_overlay rewrites the RGBA accumulator and
state.label_map directly from the polygon set in ascending label_id order.
Drops the _overlay_tracked mask-reference diff; adds a pinned-hash
snapshot test guarding the blend result.

Wave 1C — end_brush_stroke now rasterizes the target polygon into a
bbox-local scratch buffer (union of stroke + target vertex bboxes),
performs the boolean op on the crops, and commits a BrushStrokeCommand.
No stored per-region mask is consulted.

state.region_masks / state.region_areas remain allocated and written by
commands — wave 2 removes them.

+++ bacmask/core/io_manager.py
@@ -61,7 +61,7 @@ class AreaRow:
     filename: str
     region_id: int
     region_name: str
-    area_px: int
+    area_px: float  # polygon shoelace area in px² (knowledge/030)
     area_mm2: float | None  # None -> empty cell (uncalibrated)
     scale_factor: float | None  # None -> empty cell (uncalibrated)
